+++ src/rdkafka_broker.c
@@ -4212,7 +4212,8 @@ static void rd_kafka_fetch_reply_handle_partition_error (
                 err_offset = rktp->rktp_offsets.fetch_offset;
                 rktp->rktp_offsets.fetch_offset = RD_KAFKA_OFFSET_INVALID;
                 rd_kafka_offset_reset(rktp, err_offset, err,
-                                      rd_kafka_err2str(err));
+                                      "fetch failed due to requested offset "
+                                      "not available on the broker");
         }
         break;